UBUNTU: [Packaging] azure-fde: Include erofs in linux-modules instead of linux-modules-extra
BugLink: https://bugs.launchpad.net/bugs/2081613

erofs is starting to see more use as a modern alternative to squashfs.
Since squashfs is already included in linux-modules, let's include erofs
in linux-modules as well so that it can be used as a filesystem in virtual
machines with just linux-virtual installed.

Signed-off-by: Vinicius Peixoto <vinicius.peixoto@canonical.com>
Acked-by: John Cabaj <john.cabaj@canonical.com>
Acked-by: Agathe Porte <agathe.porte@canonical.com>
Acked-by: Manuel Diewald <manuel.diewald@canonical.com>
Signed-off-by: John Cabaj <john.cabaj@canonical.com>

(copied from azure)

Signed-off-by: John Cabaj <john.cabaj@canonical.com>