UBUNTU: [Packaginga] azure-fde: Don't generate list of source files
BugLink: https://bugs.launchpad.net/bugs/2104911

Generating the list of source files used for the build takes a very long
time on some architecures. Up to 30 mins. Lets not do that. We have all
that information in the debug packages and can collect it from there after
the build.

Signed-off-by: Juerg Haefliger<juerg.haefliger@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>
Acked-by: Zixing Liu <zixing.liu@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>

(copied from azure)

Signed-off-by: John Cabaj <john.cabaj@canonical.com>