UBUNTU: [Packaging] azure-fde-6.17:Revert cross compile changes
The changes made in Questing to enable cross compiling the tools do not
work with the current test compile infrastrucure. This appears to be an
issue with the package resolver used.
So at least for the time being lets change back to allow cross compiling
only for the kernel and not the tools.

Ignore: yes
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>

(copied from azure-6.17)

Signed-off-by: John Cabaj <john.cabaj@canonical.com>